fix: Delete Grid component
