--- frontend/src/components/common/Box/Box.types.ts
@@ -1,9 +1,0 @@
-import type { HTMLAttributes, ElementType, CSSProperties } from 'react';
-
-export type BoxProps = {
-  as?: ElementType;
-  display?: CSSProperties['display'];
-  flexDirection?: CSSProperties['flexDirection'];
-  justifyContent?: CSSProperties['justifyContent'];
-  alignItems?: CSSProperties['alignItems'];
-} & HTMLAttributes<HTMLDivElement>;
