--- frontend/src/components/common/Grid/Grid.types.ts
@@ -1,8 +1,0 @@
-import type { HTMLAttributes, ElementType, CSSProperties } from 'react';
-
-export type GridProps = {
-  as?: ElementType;
-  display?: CSSProperties['display'];
-  gridTemplateColumns?: CSSProperties['gridTemplateColumns'];
-  gap?: CSSProperties['gap'];
-} & HTMLAttributes<HTMLDivElement>;
