--- frontend/src/components/common/Grid/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Grid';
