+++ frontend/src/components/common/index.ts
@@ -1,3 +1,2 @@
 export { default as Box } from './Box';
 export { default as Layout } from './Layout';
-export { default as Grid } from './Grid';
